TD-50: Renombrar los archivos assertions de la carpeta labels
