+++ src/assertions/labels/rename_shared_labels_assertions.py
@@ -1,3 +1,6 @@
+from src.assertions.common_assertions import assert_empty_body
+
+
 def assert_rename_share_labels_forbidden(response):
     assert_empty_body(response, 401, "text/plain; charset=utf-8", b'Forbidden')
 
@@ -16,9 +19,3 @@ def assert_a_share_label_exists(label, share_labels):
 
 def assert_there_is_no_a_share_label(label, share_labels):
     assert str(label) not in share_labels
-
-
-def assert_empty_body(response, status_code, content_type, content=b''):
-    assert response.status_code == status_code
-    assert response.headers['Content-Type'] == content_type
-    assert response.content == content
