+++ src/assertions/labels/update_a_personal_label_assertions.py
@@ -1,5 +1,6 @@
 from src.utils.json_reader import read_a_json
 from src.assertions.schema_assertion import assert_schema
+from src.assertions.common_assertions import assert_empty_body
 
 
 def assert_update_a_label_case_one(response, label_data):
@@ -29,12 +30,8 @@ def assert_update_a_label_color_format_not_valid(response):
 
 
 def assert_update_a_label_not_found(response):
-    assert response.status_code == 404
-    assert response.headers['Content-Type'] == 'text/plain; charset=utf-8'
-    assert response.content == b'Label not found'
+    assert_empty_body(response, 404, 'text/plain; charset=utf-8', b'Label not found')
 
 
 def assert_bad_request(response, content):
-    assert response.status_code == 400
-    assert response.headers['Content-Type'] == 'text/plain; charset=utf-8'
-    assert response.content == content
+    assert_empty_body(response, 400, 'text/plain; charset=utf-8', content)
